# Improved key handling when assigning with ':='