--- a/inst/tests/tests.Rraw
+++ b/inst/tests/tests.Rraw
@@ -5738,14 +5738,14 @@ test(1396, DT[a==2, verbose=TRUE], DT[2], output="Creating new index 'a'")
 test(1397, DT[b==6, verbose=TRUE], DT[3], output="Creating new index 'b'")
 test(1398, DT[b==6, verbose=TRUE], DT[3], output="Using existing index 'b'")
 test(1399, indices(DT), c("a","b"))
-test(1400, DT[2, a:=4L, verbose=TRUE], data.table(a=c(1L,4L,3L),b=4:6), output="Dropping index 'a' due to update on 'a' (column 1)")
+test(1400, DT[2, a:=4L, verbose=TRUE], data.table(a=c(1L,4L,3L),b=4:6), output=".*Dropping index 'a' due to an update on a key column")
 test(1401, indices(DT), "b")
-test(1402, DT[,b:=NULL,verbose=TRUE], data.table(a=c(1L,4L,3L)), output="Dropping index 'b' due to delete of 'b' (column 2)")
+test(1402, DT[,b:=NULL,verbose=TRUE], data.table(a=c(1L,4L,3L)), output=".*Dropping index 'b' due to an update on a key column")
 test(1403, indices(DT), NULL)
 DT = data.table(x=1:5)
 test(1404, DT[, y := x <= 2L], data.table(x=1:5, y=c(TRUE,TRUE,FALSE,FALSE,FALSE)))
 test(1405, DT[y == TRUE, .N, verbose=TRUE], 2L, output="Creating new index")
-test(1406, DT[, y := x <= 3L, verbose=TRUE], data.table(x=1:5, y=c(TRUE,TRUE,TRUE,FALSE,FALSE)), output="Dropping index")
+test(1406, DT[, y := x <= 3L, verbose=TRUE], data.table(x=1:5, y=c(TRUE,TRUE,TRUE,FALSE,FALSE)), output=".*Dropping index")
 test(1407, DT[y == TRUE, .N], 3L)
 DT = data.table(x=1:5, y=10:6)
 test(1408, DT[x==3,verbose=TRUE], DT[3], output="Creating")
@@ -5759,14 +5759,125 @@ test(1413, DT[x==5], DT[1])
 DT = data.table(foo=1:3, bar=4:6, baz=9:7)
 setindex(DT,foo,bar,baz)
 test(1414, indices(DT), c("foo__bar__baz"))
-test(1415, DT[2,bar:=10L,verbose=TRUE], output="Dropping index 'foo__bar__baz' due to update on 'bar'")  # test middle
-test(1416, indices(DT), NULL)
+test(1415, DT[2,bar:=10L,verbose=TRUE], output=".*Shortening index 'foo__bar__baz' to 'foo' due to an update on a key column")  # test middle
+test(1416, indices(DT), 'foo')
 setindex(DT,foo,bar,baz)
-test(1417, DT[2,baz:=10L,verbose=TRUE], output="Dropping index 'foo__bar__baz' due to update on 'baz'")  # test last
+test(1417, DT[2,baz:=10L,verbose=TRUE], output=".*Shortening index 'foo__bar__baz' to 'foo__bar' due to an update on a key column")  # test last
 setindex(DT,bar,baz)
-test(1418, DT[2,c("foo","bar"):=10L,verbose=TRUE], output="Dropping index.* due to update on 'bar'")     # test 2nd to 1st
+test(1418, DT[2,c("foo","bar"):=10L,verbose=TRUE], output=".*Dropping index.* due to an update on a key column")     # test 2nd to 1st
 setindex(DT,bar,baz)
-test(1419, DT[2,c("foo","baz"):=10L,verbose=TRUE], output="Dropping index.* due to update on 'baz'")     # test 2nd to 2nd
+test(1419, DT[2,c("foo","baz"):=10L,verbose=TRUE], output=".*Shortening index 'bar__baz' to 'bar' due to an update on a key column")     # test 2nd to 2nd
+
+## testing key retainment on assign (#2372)
+DT <- data.table(x1 = c(1,1,1,1,1,2,2,2,2,2),
+                 x2 = c(1,1,2,2,2,1,1,2,2,2),
+                 x3 = c(1,2,1,1,2,1,1,1,1,2),
+                 y  = rnorm(10), 
+                 key = c("x1", "x2", "x3"))
+thisDT <- copy(DT)[2, x1 := 3]
+test(1419.1, key(thisDT), NULL)
+thisDT <- copy(DT)[2, x2 := 3]
+test(1419.2, key(thisDT), "x1")
+test(1419.3, data.table:::forderv(thisDT, c("x1")), integer(0))
+thisDT <- copy(DT)[2, x2 := 3]
+test(1419.4, key(thisDT), "x1")
+test(1419.5, data.table:::forderv(thisDT, c("x1")), integer(0))
+thisDT <- copy(DT)[3, x3 := 3]
+test(1419.6, key(thisDT), c("x1", "x2"))
+test(1419.7, data.table:::forderv(thisDT, c("x1", "x2")), integer(0))
+thisDT <- copy(DT)[3, c("x1", "x3") := .(3,3)]
+test(1419.8,key(thisDT), NULL)
+thisDT <- copy(DT)[3, c("x2", "x3") := .(3,3)]
+test(1419.9, key(thisDT), "x1")
+test(1419.10, data.table:::forderv(thisDT, c("x1")), integer(0))
+setkey(DT, NULL)
+thisDT <- copy(DT)[3, x3 := 3]
+test(1419.11, key(thisDT), NULL)
+## same tests for empty DT
+## forderv tests can be skipped for empty DT
+DT <- DT[0]
+thisDT <- copy(DT)[, x3 := 3]
+test(1419.12, key(thisDT), NULL)
+setkeyv(DT, c("x1", "x2", "x3"))
+thisDT <- copy(DT)[, x1 := 3]
+test(1419.13, key(thisDT), NULL)
+thisDT <- copy(DT)[, x2 := 3]
+test(1419.14, key(thisDT), "x1")
+thisDT <- copy(DT)[, x2 := 3]
+test(1419.15, key(thisDT), "x1")
+thisDT <- copy(DT)[, x3 := 3]
+test(1419.16, key(thisDT), c("x1", "x2"))
+thisDT <- copy(DT)[, c("x1", "x3") := .(3,3)]
+test(1419.17, key(thisDT), NULL)
+thisDT <- copy(DT)[, c("x2", "x3") := .(3,3)]
+test(1419.18, key(thisDT), "x1")
+
+## testing secondary key retainment on assign (#2372)
+DT <- data.table(a = c(1,1,1,2,1,2,2,2,2,2),
+                 aaa = c(2,1,2,2,2,1,1,2,2,2),
+                 b = c(1,2,1,1,2,1,1,1,1,2),
+                 ab  = rnorm(10))
+allIndicesValid <- function(DT){
+  ## checks that the order of all indices is correct
+  for(idx in seq_along(indices(DT))){
+    index <- attr(attr(DT, "index"), paste0("__", indices(DT)[idx], collapse = ""))
+    if(!length(index)) index <- seq_len(nrow(DT))
+    if(length(data.table:::forderv(DT[index],
+                                   indices(DT, vectors = TRUE)[[idx]]))){
+      ## index is not properly ordered
+      return(FALSE)
+    }
+    if(any(duplicated(names(attributes(attr(DT, "index")))))){
+      ## index names are not unique
+      return(FALSE)
+    }
+  }
+  return(TRUE)
+}
+test(1419.19, indices(copy(DT)[1, a:=1]), NULL)
+setindex(DT, a)
+setindex(DT, a, aaa)
+setindex(DT, ab, aaa)
+setindex(DT)
+test(1419.21, allIndicesValid(DT), TRUE)
+thisDT <- copy(DT)[1, a:=1][, aaa := 1][, ab := 1]
+test(1419.22, indices(thisDT), NULL)
+test(1419.23, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, b := 2]
+test(1419.24, indices(thisDT), c("a", "a__aaa", "ab__aaa"))
+test(1419.25, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, ab := 2]
+test(1419.26, indices(thisDT), c("a", "a__aaa", "a__aaa__b"))
+test(1419.27, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, aaa := 2]
+test(1419.28, indices(thisDT), c("a", "ab"))
+test(1419.29, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, c("aaa", "b") := 2]
+test(1419.30, indices(thisDT), c("a", "ab"))
+test(1419.31, allIndicesValid(thisDT), TRUE)
+## same on empty DT
+DT <- DT[0]
+setindex(DT, a)
+setindex(DT, a, aaa)
+setindex(DT, ab, aaa)
+setindex(DT)
+test(1419.32, allIndicesValid(DT), TRUE)
+thisDT <- copy(DT)[, a:=1][, aaa := 1][, ab := 1]
+test(1419.33, indices(thisDT), NULL)
+test(1419.34, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, b := 2]
+test(1419.35, indices(thisDT), c("a", "a__aaa", "ab__aaa"))
+test(1419.36, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, ab := 2]
+test(1419.37, indices(thisDT), c("a", "a__aaa", "a__aaa__b"))
+test(1419.38, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, aaa := 2]
+test(1419.39, indices(thisDT), c("a", "ab"))
+test(1419.40, allIndicesValid(thisDT), TRUE)
+thisDT <- copy(DT)[, c("aaa", "b") := 2]
+test(1419.41, indices(thisDT), c("a", "ab"))
+test(1419.42, allIndicesValid(thisDT), TRUE)
+
 
 # setnames updates secondary key
 DT = data.table(a=1:5,b=10:6)