@mattdowle 2017-09-27
Sorry - I guess I broke it by removing `shift` alias.  I didn't think that was needed as `shift` is exported. I'll fix.

I've added you as a project member so that you can create branches inside the main repo if you'd like.  This will mean I can pull, test and push to your branch, rather than me using the online editor as I'm doing now.